Move actual value transformation into separate file
